Introduce 'Smart Decimation' that preserves shape keys
The Decimate modifier cannot be applied to any mesh which has shape
keys, but the Decimate mesh operator can. Applying the Decimate mesh
operator causes the shape keys to all get jumbled up, but notably
jumbled up in __the same way__, so we can exploit this by creating a
duplicate Basis key that we let get messed up, and use it as a reference
to un-jumble the other shape keys.