MMD shader fix Auto Bake
fixed MMD shaders with ambient occlusion and diffuse color to bake into a new texture and apply it to the principled BSDF during the BSDF adding stage during fix model. This is so things like MMD hair textures with color changing capabilities on a model are supported now. (ex: White hair texture with green diffuse multiply)