Change "Fix Model" To "Standardize Model"
To fix confusion people have in the discord on "Why isn't this """fixing""" my model?" maybe a better term for this button would be standardize. This is because standardizing is what it does. It makes the model bones have standard names (The ones cats uses) and it makes everything work for VRChat.

Sure it does "fix" the model, but this can be taken in the wrong context (And has on almost every occasion that someone complained about the button) because it doesn't "fix" any issue thrown, which is what makes people think it's a """magic button""" because they think it can fix anything thrown at it.

I hope this fixes the problem of people misinterpreting it's usage.

**name open to changes** but a name other than "Fix Model" that is reflective of it's actual function and would help sort out confusion would be much appreciated

Thanks!